move existing fixers to transformers
